made changes

+++ Statistics.Test/StatsUnitTest.cs
@@ -13,7 +13,7 @@ namespace Statistics.Test
             var StatisticsComputer = new StatisticsComputer();
             var computedStats = StatisticsComputer.CalculateStatistics(
                 new List<___>{1.5, 8.9, 3.2, 4.5});
-            double epsilon = 0.001F;
+            float epsilon = 0.001F;
             Assert.True(Math.Abs(StatisticsComputer._avg - 4.525) <= epsilon);
             Assert.True(Math.Abs(StatisticsComputer._max - 8.9) <= epsilon);
             Assert.True(Math.Abs(StatisticsComputer._min - 1.5) <= epsilon);
@@ -42,7 +42,7 @@ namespace Statistics.Test
             var ledAlert = new LedAlerter();
             IAlerter[] alerters = {emailAlert, ledAlert};
 
-            const double maxThreshold = 10.2;
+            const float maxThreshold = 10.2;
             var statsAlerter = new StatisticsAlerter(maxThreshold, alerters);
             statsAlerter.checkAndAlert(new List<___>{0.2, 11.9, 4.3, 8.5});
